requirements.txt
holoviews==1.17.0
matplotlib==3.5.1
numpy==1.22.4
pandas==1.5.3
openpyxl==3.1.2
plotly==5.5.0
kaleido==0.2.1
streamlit==1.24.0
Pillow==9.5.0